@@ -5,8 +5,8 @@ import android.view.View
 import android.view.ViewGroup
 import com.jmoraes.componentizationsample.R
 import com.jmoraes.componentizationsample.basic.eventTypes.UserInteractionEvent
-import com.netflix.arch.EventBusFactory
-import com.netflix.arch.UIView
+import com.netflix.componentizationV1.EventBusFactory
+import com.netflix.componentizationV1.UIView
 
 class SuccessView(container: ViewGroup, eventBusFactory: EventBusFactory) :
     UIView<UserInteractionEvent>(container) {
