@@ -4,7 +4,7 @@ import android.arch.lifecycle.LifecycleOwner
 import android.view.ViewGroup
 import com.jmoraes.componentizationsample.basic.eventTypes.ScreenStateEvent
 import com.jmoraes.componentizationsample.basic.components.uiViews.ErrorView
-import com.netflix.arch.EventBusFactory
+import com.netflix.componentizationV1.EventBusFactory
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import org.junit.Before
