@@ -2,9 +2,7 @@ package com.jmoraes.componentizationsample.player.components
 
 import android.arch.lifecycle.LifecycleOwner
 import android.view.ViewGroup
-import com.jmoraes.componentizationsample.basic.eventTypes.ScreenStateEvent
-import com.jmoraes.componentizationsample.basic.components.uiViews.ErrorView
-import com.netflix.arch.EventBusFactory
+import com.netflix.componentizationV1.EventBusFactory
 import com.netflix.elfo.components.PrimaryControlsUIView
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
